Support If-Modified-Since header when sending Etag responses
The `fresh` function will return `false` if the request headers include `If-Modified-Since` but the response headers do not include `Last-Modified`. Next.js is not passing the `Last-Modified` header to `fresh` so this causes it to return `false`.

Another option would to remove the `If-Modified-Since` header from the headers passed to `fresh`,